+++ ios/ReactNativeAudioToolkit/ReactNativeAudioToolkit/AudioPlayer.m
@@ -137,8 +137,9 @@ RCT_EXPORT_METHOD(prepare:(nonnull NSNumber*)playerId
             avAudioSessionCategory = AVAudioSessionCategorySoloAmbient;
             break;
     }
+    NSNumber *mixWithOthers = [options objectForKey:@"mixWithOthers"];
     NSError *error = nil;
-    [[AVAudioSession sharedInstance] setCategory: avAudioSessionCategory error: &error];
+    [[AVAudioSession sharedInstance] setCategory: avAudioSessionCategory withOptions: mixWithOthers ? AVAudioSessionCategoryOptionMixWithOthers : 0 error: &error];
     if (error) {
         NSDictionary* dict = [Helpers errObjWithCode:@"preparefail"
                                          withMessage:@"Failed to set audio session category."];